fix: split home page into multiple files
