+++ composeApp/src/androidMain/kotlin/com/eren76/mangly/composables/screens/Settings.kt
@@ -53,6 +53,7 @@ import androidx.compose.ui.unit.dp
 import androidx.core.content.edit
 import androidx.navigation.NavHostController
 import com.eren76.mangly.Constants
+import com.eren76.mangly.composables.screens.home.HomeSorting
 import com.eren76.mangly.composables.screens.readviewer.ReaderModePrefs
 import com.eren76.mangly.composables.screens.readviewer.ReaderModeType
 import com.eren76.mangly.themes.setAppTheme
